Backend/requirements.txt
discord.py
requests
fastapi
uvicorn